Partway to Generic SynthItems
Quite a ways toward a SynthItem base class that includes generic
commands and a RingBuffer, so that it could be the only class wrapped in
Qt (or another GUI framework) and hold any SynthItem derived class.
However, it is messy and I'm not sold on the idea now.

Going to push, then branch from an earlier version, where each class was
wrapped individually.