QtAudifier
Simple wrapping of Audifier for Qt implemented.

QtOscillator and QtAudifier stripped down, now only implement funtions
not available in the base.  Realised I was confuins the Qt wrapper
classes with the Son classes, in that I kept acting as if the wrapper
would require different overrides of the base class core functions (i.e.
process, addchild), but really , it is the son::SyntItem classes that
have the different implementations.  They are all called from ther
wrapper the same way.