Modulator mostly implemented.
At this point essentially an Oscillator with only one generator.  Also
changed so that addChild in SynthItems and in QML automatically calls
addParent on the child and passes self as argument.

AddChild now returns bool.  It checks the type of the child and rejects
inappropriate children, after checking against a list.

Other changes here and there to make it all work, minor bug fixes.