SynthItem.qml delete fixes
Fixed bug where SynthItems connected to OUT (root) were not receiving
the parent pointer under the new system.  This is because OUT has it's
own addChild function, which was not reciprocally calling addParent.

This is another thing that makes me realize, the cpp library classes
SynthItem and SynthGraph should be derived from the same base.  All
adding and removing of children/parents should be handled the same way.
Other reasons too.